remove unnecessary code
