+++ src/redux/actionsTypes/notifications.ts
@@ -1,2 +1,1 @@
 export const ADD_NOTIFICATION = 'ADD_NOTIFICATION';
-export const REMOVE_NOTIFICATION = 'REMOVE_NOTIFICATION';
